+++ R/params.R
@@ -4,9 +4,9 @@ total_effect_parameters <- list(
 )
 
 ria_parameters <- list(
-	c(i = "data_1zp", j = "data_1", k = "data_0", l = "data_0"),
-	c(i = "data_0zp", j = "data_0", k = "data_0", l = "data_0"),
-	c(i = "data_1zp", j = "data_1", k = "data_1", l = "data_1")
+	c(i = "data_1lp", j = "data_1", k = "data_0", l = "data_0"),
+	c(i = "data_0lp", j = "data_0", k = "data_0", l = "data_0"),
+	c(i = "data_1lp", j = "data_1", k = "data_1", l = "data_1")
 )
 
 estimation_parameters <- list(