+++ R/phi.R
@@ -35,7 +35,7 @@ phi_n_alpha <- function(train, valid, vars, architecture, params, control) {
 	alpha1 <- Alpha(
 		train = train,
 		valid = valid,
-		vars = na.omit(c(vars@A, vars@W)),
+		vars = na.omit(c(vars@D, vars@C)),
 		architecture = architecture,
 		.f = .f1,
 		weights = NULL,
@@ -45,7 +45,7 @@ phi_n_alpha <- function(train, valid, vars, architecture, params, control) {
 	alpha2 <- Alpha(
 		train = train,
 		valid = valid,
-		vars = na.omit(c(vars@A, vars@Z, vars@W)),
+		vars = na.omit(c(vars@D, vars@L, vars@C)),
 		architecture = architecture,
 		.f = .f2,
 		weights = alpha1$train,
@@ -55,7 +55,7 @@ phi_n_alpha <- function(train, valid, vars, architecture, params, control) {
 	alpha3 <- Alpha(
 		train = train,
 		valid = valid,
-		vars = na.omit(c(vars@A, vars@C, vars@M, vars@Z, vars@W)),
+		vars = na.omit(c(vars@D, vars@observed, vars@M, vars@L, vars@C)),
 		architecture = architecture,
 		.f = .f3,
 		weights = alpha2$train,
@@ -83,7 +83,7 @@ estimate_phi_r_alpha <- function(cd, folds, params, nn_module, control) {
 			)
 
 			names(alpha_rs[[i]]) <-
-				gsub("zp", "", unlist(lapply(
+				gsub("lp", "", unlist(lapply(
 					params$randomized, \(x) paste0(gsub("data_", "", x), collapse = "")
 				)))
 
@@ -109,7 +109,7 @@ phi_r_alpha <- function(train, valid, vars, architecture, params, control) {
 	alpha1 <- Alpha(
 		train = train,
 		valid = valid,
-		vars = na.omit(c(vars@A, vars@W)),
+		vars = na.omit(c(vars@D, vars@C)),
 		architecture = architecture,
 		.f = .f1,
 		weights = NULL,
@@ -119,7 +119,7 @@ phi_r_alpha <- function(train, valid, vars, architecture, params, control) {
 	alpha2 <- Alpha(
 		train = train,
 		valid = valid,
-		vars = na.omit(c(vars@A, vars@Z, vars@W)),
+		vars = na.omit(c(vars@D, vars@L, vars@C)),
 		architecture = architecture,
 		.f = .f2,
 		weights = alpha1$train,
@@ -129,7 +129,7 @@ phi_r_alpha <- function(train, valid, vars, architecture, params, control) {
 	alpha3 <- Alpha(
 		train = train,
 		valid = valid,
-		vars = na.omit(c(vars@A, vars@M, vars@W)),
+		vars = na.omit(c(vars@D, vars@M, vars@C)),
 		architecture = architecture,
 		.f = .f3,
 		weights = alpha2$train,
@@ -139,7 +139,7 @@ phi_r_alpha <- function(train, valid, vars, architecture, params, control) {
 	alpha4 <- Alpha(
 		train = train,
 		valid = valid,
-		vars = na.omit(c(vars@A, vars@C, vars@Z, vars@M, vars@W)),
+		vars = na.omit(c(vars@D, vars@observed, vars@L, vars@M, vars@C)),
 		architecture = architecture,
 		.f = .f4,
 		weights = alpha3$train,