+++ R/ria.test_vars.R
@@ -5,19 +5,19 @@ NULL
 
 ria.test_vars <- S7::new_class("ria.test_vars",
   properties = list(
-    A = S7::class_character,
+    D = S7::class_character,
     Y = S7::class_character,
     M = S7::class_character,
-    Z = S7::class_character,
-    W = S7::class_character,
-    C = S7::new_property(class = S7::class_character, default = NA_character_),
+    L = S7::class_character,
+    C = S7::class_character,
+    observed = S7::new_property(class = S7::class_character, default = NA_character_),
     id = S7::new_property(class = S7::class_character, default = NA_character_)
   ),
   validator = function(self) {
      if (length(self@Y) != 1) {
       "self@outcome must be length 1"
-    } else if (length(self@C) != 1) {
-    	"self@cens must be length 1"
+    } else if (length(self@observed) != 1) {
+      "self@observed must be length 1"
     } else if (length(self@id) != 1) {
     	"self@id must be length 1"
     }